- Fix Ctrl + wheel shortcut - Fix zoomUp at the end of the bar - add a stretch on the right

+++ src/webview.h
@@ -74,7 +74,7 @@ private slots:
 
 signals:
     void loadUrl(const KUrl &, const Rekonq::OpenType &);
-    void zoomChanged(qreal);
+    void zoomChanged(int);
 
 private:
     QPoint _mousePos;
